Removed logic that generates include paths in macros
This was confusing distributed builds.
Also added type name to the JPH_SHADER_CONSTANTS_END, JPH_SHADER_BIND_END and JPH_SHADER_STRUCT_END to make it easier to integrate in other engines
