+++ Jolt/Shaders/ShaderCore.h
@@ -25,10 +25,10 @@
 
 	#define JPH_SHADER_CONSTANTS_BEGIN(type, name)	struct type {
 	#define JPH_SHADER_CONSTANTS_MEMBER(type, name)	type c##name;
-	#define JPH_SHADER_CONSTANTS_END				};
+	#define JPH_SHADER_CONSTANTS_END(type)			};
 
 	#define JPH_SHADER_BIND_BEGIN(name)
-	#define JPH_SHADER_BIND_END
+	#define JPH_SHADER_BIND_END(name)
 	#define JPH_SHADER_BIND_BUFFER(type, name)
 	#define JPH_SHADER_BIND_RW_BUFFER(type, name)
 
@@ -53,7 +53,7 @@
 
 	#define JPH_SHADER_CONSTANTS_BEGIN(type, name)	cbuffer name {
 	#define JPH_SHADER_CONSTANTS_MEMBER(type, name)	type c##name;
-	#define JPH_SHADER_CONSTANTS_END				};
+	#define JPH_SHADER_CONSTANTS_END(type)			};
 
 	#define JPH_SHADER_FUNCTION_BEGIN(return_type, name, group_size_x, group_size_y, group_size_z) \
 		[numthreads(group_size_x, group_size_y, group_size_z)] \
@@ -65,7 +65,7 @@
 	#define JPH_SHADER_RW_BUFFER(type)				RWStructuredBuffer<type>
 
 	#define JPH_SHADER_BIND_BEGIN(name)
-	#define JPH_SHADER_BIND_END
+	#define JPH_SHADER_BIND_END(name)
 	#define JPH_SHADER_BIND_BUFFER(type, name)		JPH_SHADER_BUFFER(type) name;
 	#define JPH_SHADER_BIND_RW_BUFFER(type, name)	JPH_SHADER_RW_BUFFER(type) name;
 
@@ -74,7 +74,7 @@
 
 #define JPH_SHADER_STRUCT_BEGIN(name)				struct name {
 #define JPH_SHADER_STRUCT_MEMBER(type, name)		type m##name;
-#define JPH_SHADER_STRUCT_END						};
+#define JPH_SHADER_STRUCT_END(name)					};
 
 #define JPH_IN(type)								in type
 #define JPH_OUT(type)								out type
