+++ Jolt/Shaders/TestComputeBindings.h
@@ -2,6 +2,8 @@
 // SPDX-FileCopyrightText: 2025 Jorrit Rouwe
 // SPDX-License-Identifier: MIT
 
+#include "ShaderCore.h"
+
 JPH_SHADER_CONSTANT(int, cTestComputeGroupSize, 64)
 
 JPH_SHADER_CONSTANTS_BEGIN(TestComputeContext, gContext)
@@ -10,10 +12,10 @@ JPH_SHADER_CONSTANTS_BEGIN(TestComputeContext, gContext)
 	JPH_SHADER_CONSTANTS_MEMBER(JPH_float3,		Float3Value2)
 	JPH_SHADER_CONSTANTS_MEMBER(JPH_uint,		UIntValue2)
 	JPH_SHADER_CONSTANTS_MEMBER(JPH_uint,		NumElements)
-JPH_SHADER_CONSTANTS_END
+JPH_SHADER_CONSTANTS_END(TestComputeContext)
 
 JPH_SHADER_BIND_BEGIN(JPH_TestCompute)
 	JPH_SHADER_BIND_BUFFER(JPH_uint, gUploadData)
 	JPH_SHADER_BIND_BUFFER(JPH_uint, gOptionalData)
 	JPH_SHADER_BIND_RW_BUFFER(JPH_uint, gData)
-JPH_SHADER_BIND_END
+JPH_SHADER_BIND_END(JPH_TestCompute)
