+++ Jolt/Shaders/TestComputeWrapper.cpp
@@ -5,5 +5,8 @@
 #include <Jolt/Jolt.h>
 
 #define JPH_SHADER_NAME TestCompute
-#include "ShaderWrapperCreator.h"
-#undef JPH_SHADER_NAME
+#include <Jolt/Compute/CPU/WrapShaderBegin.h>
+#include "TestCompute.hlsl"
+#include <Jolt/Compute/CPU/WrapShaderBindings.h>
+#include "TestComputeBindings.h"
+#include <Jolt/Compute/CPU/WrapShaderEnd.h>
